[Bugfix] Fixed date picker tests in Add Contract Form
What:
Ran into a edge case that caused the start and
end date tests in the Add contract form to fail.

When the current date is the start of the month
the new start  or end dates would be in the
previous month. The datepicker only
shows the default month, hence the values
were not displayed on the screen.

Why:
We need the tests to pass consistently for
CI builds